+++ lib/domain/workout/entities/month_workout.dart
@@ -6,7 +6,6 @@ import 'package:power_progress/domain/workout/entities/accumulation_workout.dart
 import 'package:power_progress/domain/workout/entities/deload_workout.dart';
 import 'package:power_progress/domain/workout/entities/intensification_workout.dart';
 import 'package:power_progress/domain/workout/entities/realization_workout.dart';
-import 'package:power_progress/domain/workout/entities/workout_failure.dart';
 
 class MonthWorkout {
   final Month month;
@@ -15,6 +14,8 @@ class MonthWorkout {
   final IntensificationWorkout intensificationWorkout;
   final RealizationWorkout realizationWorkout;
   final DeloadWorkout deloadWorkout;
+  final bool isPreviousDeloadDone;
+  final bool isNextAccumulationDone;
 
   MonthWorkout({
     @required this.month,
@@ -23,35 +24,27 @@ class MonthWorkout {
     @required this.intensificationWorkout,
     @required this.realizationWorkout,
     @required this.deloadWorkout,
+    @required this.isPreviousDeloadDone,
+    @required this.isNextAccumulationDone,
   });
 
   bool validatable(WeekEnum week) {
-    switch (week) {
-      case WeekEnum.accumulation:
-        return true;
-      case WeekEnum.intensification:
-        return accumulationWorkout.isDone;
-      case WeekEnum.realization:
-        return intensificationWorkout.isDone;
-      case WeekEnum.deload:
-        return realizationWorkout.isDone;
-      default:
-        throw const UnexpectedError();
-    }
+    return week.when(
+      // only if it's the first week or last week's deload is done
+      accumulation: () => isPreviousDeloadDone,
+      intensification: () => accumulationWorkout.isDone,
+      realization: () => intensificationWorkout.isDone,
+      deload: () => realizationWorkout.isDone,
+    );
   }
 
   bool invalidatable(WeekEnum week) {
-    switch (week) {
-      case WeekEnum.accumulation:
-        return !intensificationWorkout.isDone;
-      case WeekEnum.intensification:
-        return !realizationWorkout.isDone;
-      case WeekEnum.realization:
-        return !deloadWorkout.isDone;
-      case WeekEnum.deload:
-        return true;
-      default:
-        throw const UnexpectedError();
-    }
+    return week.when(
+      accumulation: () => !intensificationWorkout.isDone,
+      intensification: () => !realizationWorkout.isDone,
+      realization: () => !deloadWorkout.isDone,
+      // only if next week's accumulation is not done
+      deload: () => !isNextAccumulationDone,
+    );
   }
 }