+++ lib/domain/workout/usecases/mark_workout_undone.dart
@@ -30,18 +30,20 @@ class MarkWorkoutUndone implements UseCase<Unit, WorkoutFailure, MarkWorkoutUndo
     await updateExerciseNextWeek(
       UpdateExerciseNextWeekParams(
         exerciseId: params.exerciseId,
-        nextWeek: params.week.previous(),
+        nextWeek: params.week,
       ),
     );
 
-    if (params.week == WeekEnum.deload) {
-      await updateExerciseNextMonth(
+    // if it's the deload week that is being marked undone then the current month become the current on again and not the next one.
+    await params.week.maybeWhen(
+      deload: () async => updateExerciseNextMonth(
         UpdateExerciseNextMonthParams(
           exerciseId: params.exerciseId,
-          nextMonth: params.month.previous,
+          nextMonth: params.month,
         ),
-      );
-    }
+      ),
+      orElse: () {},
+    );
 
     return result;
   }